requirements.txt
Click==7.0
Flask==1.1.1
gunicorn==19.10.0
itsdangerous==1.1.0
Jinja2==2.10.3
MarkupSafe==1.1.1
Werkzeug==0.16.0
Flask-RESTful==0.3.7 
pandas==0.25.3
camelot-py==0.7.3
